+++ main.js
@@ -1,22 +1,22 @@
 #!/usr/bin/env node
-/*
- *
- * Copyright 2013 Anis Kadri
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
+/**
+    Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+*/
 
 // copyright (c) 2013 Andrew Lunny, Adobe Systems
 var path = require('path')
@@ -28,7 +28,7 @@ var path = require('path')
     , cordova_lib = require('cordova-lib')
     , plugman = cordova_lib.plugman;
 
-var known_opts = { 'platform' : [ 'ios', 'android', 'amazon-fireos', 'blackberry10', 'wp7', 'wp8' , 'windows8', 'firefoxos' ]
+var known_opts = { 'platform' : [ 'ios', 'android', 'amazon-fireos', 'blackberry10', 'wp8' , 'windows8', 'firefoxos' ]
         , 'project' : path
         , 'plugin' : [String, path, url, Array]
         , 'version' : Boolean
@@ -40,6 +40,7 @@ var known_opts = { 'platform' : [ 'ios', 'android', 'amazon-fireos', 'blackberry
         , 'variable' : Array
         , 'www': path
         , 'searchpath' : [path, Array]
+        , 'browserify': Boolean
 }, shortHands = { 'var' : ['--variable'], 'v': ['--version'], 'h': ['--help'] };
 
 var cli_opts = nopt(known_opts, shortHands);
@@ -55,7 +56,7 @@ if (cli_opts.plugins_dir || cli_opts.project) {
 
 process.on('uncaughtException', function(error) {
     if (cli_opts.debug) {
-        console.error(error.stack);
+        console.error(error.message, error.stack);
     } else {
         console.error(error.message);
     }